# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Gym CRM REST API
  version: 1.0.0
  description: |
    REST API for the Gym CRM system.

    Public endpoints are limited to trainee registration, trainer registration,
    and login. Every other endpoint requires a Bearer JWT.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/trainers/register:
    post:
      security: []
      tags:
        - Trainers
      summary: Register a trainer
      operationId: registerTrainer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerRegistrationRequest'
      responses:
        '201':
          description: Trainer registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/trainers/profile:
    get:
      tags:
        - Trainers
      summary: Get the authenticated trainer's profile
      operationId: getTrainerProfile
      responses:
        '200':
          description: Trainer profile returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Trainers
      summary: Update the authenticated trainer's profile
      operationId: updateTrainerProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerUpdateRequest'
      responses:
        '200':
          description: Trainer profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainerProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/trainers/status:
    patch:
      tags:
        - Trainers
      summary: Activate or deactivate the authenticated trainer
      operationId: updateTrainerStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivationRequest'
      responses:
        '200':
          description: Trainer status updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/trainees/trainings:
    get:
      tags:
        - Trainees
      summary: Get trainings for the authenticated trainee
      operationId: getTraineeTrainings
      parameters:
        - name: fromDate
          in: query
          required: false
          description: Include trainings on or after this date
          schema:
            type: string
            format: date
          example: '2026-01-01'
        - name: toDate
          in: query
          required: false
          description: Include trainings on or before this date
          schema:
            type: string
            format: date
          example: '2026-12-31'
        - name: trainerUsername
          in: query
          required: false
          description: Filter by trainer username
          schema:
            type: string
          example: Mike.Brown
        - name: trainingTypeName
          in: query
          required: false
          description: Filter by training type
          schema:
            type: string
          example: Fitness
      responses:
        '200':
          description: Trainee trainings returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TraineeTrainingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/trainers/trainings:
    get:
      tags:
        - Trainers
      summary: Get trainings for the authenticated trainer
      operationId: getTrainerTrainings
      parameters:
        - name: fromDate
          in: query
          required: false
          description: Include trainings on or after this date
          schema:
            type: string
            format: date
          example: '2026-01-01'
        - name: toDate
          in: query
          required: false
          description: Include trainings on or before this date
          schema:
            type: string
            format: date
          example: '2026-12-31'
        - name: traineeUsername
          in: query
          required: false
          description: Filter by trainee username
          schema:
            type: string
          example: John.Smith
      responses:
        '200':
          description: Trainer trainings returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainerTrainingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/trainings:
    post:
      tags:
        - Trainings
      summary: Add a training as the authenticated trainer
      operationId: addTraining
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTrainingRequest'
      responses:
        '200':
          description: Training added successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/training-types:
    get:
      tags:
        - Trainings
      summary: Get all training types
      operationId: getTrainingTypes
      responses:
        '200':
          description: Training types returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrainingTypeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, or revoked bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource is already in the requested state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    TrainerRegistrationRequest:
      type: object
      required:
        - firstName
        - lastName
        - specialization
      properties:
        firstName:
          type: string
          example: Mike
        lastName:
          type: string
          example: Brown
        specialization:
          type: string
          example: Fitness
    RegistrationResponse:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: John.Smith
        password:
          type: string
          format: password
          readOnly: true
          description: Generated raw password returned only during registration
          example: pass123456
...
    TrainerUpdateRequest:
      type: object
      required:
        - firstName
        - lastName
        - active
      properties:
        firstName:
          type: string
          example: Mike
        lastName:
          type: string
          example: Brown
        active:
          type: boolean
          example: true
    ActivationRequest:
      type: object
      required:
        - active
      properties:
        active:
          type: boolean
          example: true
...
    AddTrainingRequest:
      type: object
      required:
        - traineeUsername
        - trainingName
        - trainingDate
        - trainingDuration
      properties:
        traineeUsername:
          type: string
          example: John.Smith
        trainingName:
          type: string
          example: Morning Training
        trainingDate:
          type: string
          format: date
          example: '2026-05-10'
        trainingDuration:
          type: integer
          format: int32
          minimum: 1
          example: 60
...
    TrainerProfileResponse:
      type: object
      required:
        - username
        - firstName
        - lastName
        - specialization
        - active
        - trainees
      properties:
        username:
          type: string
          example: Mike.Brown
        firstName:
          type: string
          example: Mike
        lastName:
          type: string
          example: Brown
        specialization:
          type: string
          example: Fitness
        active:
          type: boolean
          example: true
        trainees:
          type: array
          items:
            $ref: '#/components/schemas/TraineeShortResponse'
...
    TraineeShortResponse:
      type: object
      required:
        - traineeUsername
        - traineeFirstName
        - traineeLastName
      properties:
        traineeUsername:
          type: string
          example: John.Smith
        traineeFirstName:
          type: string
          example: John
        traineeLastName:
          type: string
          example: Smith
    TraineeTrainingResponse:
      type: object
      required:
        - trainingName
        - trainingDate
        - trainingType
        - trainingDuration
        - trainerName
      properties:
        trainingName:
          type: string
          example: Morning Training
        trainingDate:
          type: string
          format: date
          example: '2026-05-10'
        trainingType:
          type: string
          example: Fitness
        trainingDuration:
          type: integer
          format: int32
          example: 60
        trainerName:
          type: string
          example: Mike Brown
    TrainerTrainingResponse:
      type: object
      required:
        - trainingName
        - trainingDate
        - trainingType
        - trainingDuration
        - traineeName
      properties:
        trainingName:
          type: string
          example: Morning Training
        trainingDate:
          type: string
          format: date
          example: '2026-05-10'
        trainingType:
          type: string
          example: Fitness
        trainingDuration:
          type: integer
          format: int32
          example: 60
        traineeName:
          type: string
          example: John Smith
    TrainingTypeResponse:
      type: object
      required:
        - id
        - trainingTypeName
      properties:
        id:
          type: integer
          format: int64
          example: 1
        trainingTypeName:
          type: string
          example: Fitness
    ErrorResponse:
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
          format: int32
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Missing or invalid bearer token